miaDash utilities

Helper functions to support the app functionality.

.import_datasets(selection)

.update_tse(tse, fun, fun.args)

.print_message(..., title = "Invalid input:")

.set_optarg(item, loader = NULL, alternative = NULL, ...)

.check_formula(form, tse)

Arguments

selection

Numeric vector. A list of indices for the mia datasets to return.

tse

a TreeSummarizedExperiment object.

fun

Function scalar. Function to apply to tse.

fun.args

Named list. A list of arguments to pass to fun.

...

Either a series of strings to form the message returned by .print_message or named arguments for loader.

title

Character scalar. The title of the error message to print. (Default: "Invalid input:")

item

Character scalar. The file path from which the object should be loaded.

loader

Function scalar. The function to load item. (Default: NULL)

alternative

an alternative output to return when at least one item and loader are not defined. (Default: NULL)

form

Character scalar The formula to be checked.

Value

.import_datasets: returns the list of available mia datasets.

.update_tse: returns a TreeSE object after applying fun with arguments fun.args. Eventual messages and errors are also printed.

.print_message: returns a modalDialog with the error message specified with ... and titled title.

.set_optarg: returns an optional element for a TreeSE constructor. The output is either an object located at the file path item and loaded with loader or alternative.

.check_formula: returns TRUE or FALSE depending whether or not all variables included in form are present in se colData.